Move ArchiveFormSchema to filing directory and rename it.
So it can be used by the z3c.form and in a later commit by the API
transition extender.
