+++ opengever/dossier/tests/test_archiver.py
@@ -10,23 +10,23 @@ from opengever.base.behaviors.changed import IChanged
 from opengever.core.testing import COMPONENT_UNIT_TESTING
 from opengever.document.behaviors.metadata import IDocumentMetadata
 from opengever.dossier.archive import Archiver
-from opengever.dossier.archive import EnddateValidator
-from opengever.dossier.archive import get_filing_actions
-from opengever.dossier.archive import METHOD_FILING
-from opengever.dossier.archive import METHOD_RESOLVING
-from opengever.dossier.archive import METHOD_RESOLVING_AND_FILING
-from opengever.dossier.archive import METHOD_RESOLVING_EXISTING_FILING
-from opengever.dossier.archive import MissingValue
-from opengever.dossier.archive import ONLY_NUMBER
-from opengever.dossier.archive import ONLY_RESOLVE
-from opengever.dossier.archive import RESOLVE_AND_NUMBER
-from opengever.dossier.archive import RESOLVE_WITH_EXISTING_NUMBER
-from opengever.dossier.archive import RESOLVE_WITH_NEW_NUMBER
-from opengever.dossier.archive import valid_filing_year
 from opengever.dossier.behaviors.dossier import IDossier
 from opengever.dossier.behaviors.dossier import IDossierMarker
 from opengever.dossier.behaviors.filing import IFilingNumber
 from opengever.dossier.behaviors.filing import IFilingNumberMarker
+from opengever.dossier.filing.form import EnddateValidator
+from opengever.dossier.filing.form import get_filing_actions
+from opengever.dossier.filing.form import METHOD_FILING
+from opengever.dossier.filing.form import METHOD_RESOLVING
+from opengever.dossier.filing.form import METHOD_RESOLVING_AND_FILING
+from opengever.dossier.filing.form import METHOD_RESOLVING_EXISTING_FILING
+from opengever.dossier.filing.form import MissingValue
+from opengever.dossier.filing.form import ONLY_NUMBER
+from opengever.dossier.filing.form import ONLY_RESOLVE
+from opengever.dossier.filing.form import RESOLVE_AND_NUMBER
+from opengever.dossier.filing.form import RESOLVE_WITH_EXISTING_NUMBER
+from opengever.dossier.filing.form import RESOLVE_WITH_NEW_NUMBER
+from opengever.dossier.filing.form import valid_filing_year
 from opengever.dossier.interfaces import IDossierArchiver
 from opengever.testing import IntegrationTestCase
 from zope.interface import Invalid
